replace main activity code to fragment
