+++ app/src/main/java/ru/dw/gbnotes/App.java
@@ -1,17 +1,23 @@
 package ru.dw.gbnotes;
 
 import android.app.Application;
-import android.content.Context;
 
 import ru.dw.gbnotes.data.Repository;
 
 
 public class App extends Application {
+    private static App sInstance = null;
     private final Repository repository = new Repository();
 
+    @Override
+    public void onCreate() {
+        super.onCreate();
+        sInstance = this;
+    }
 
-    public static App get(Context context) {
-        return (App) context.getApplicationContext();
+
+    public static App get() {
+        return sInstance;
     }
 
     public Repository getRepository() {
