+++ app/src/main/java/ru/dw/gbnotes/ui/NoteActivity.java
@@ -28,7 +28,7 @@ public class NoteActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_note);
-        repository = App.get(this).getRepository();
+        repository = App.get().getRepository();
         initView();
         newOrUpNone();
     }
