+++ app/src/main/res/layout/fragment_note_list.xml
@@ -1,0 +1,27 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.coordinatorlayout.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    tools:context=".ui.MainActivity">
+
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/fragment_note_list_ecycler_view"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        tools:listitem="@layout/item_notes"
+        />
+    <com.google.android.material.floatingactionbutton.FloatingActionButton
+        android:id="@+id/fragment_note_list_fab"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="bottom|end"
+        android:layout_marginEnd="@dimen/double_indent"
+        android:layout_marginBottom="@dimen/double_indent"
+        android:tintMode="@color/background_bottom"
+        android:backgroundTint="@color/background_bottom"
+        app:srcCompat="@drawable/ic_add"
+        android:contentDescription="@string/content_description_fab" />
+
+</androidx.coordinatorlayout.widget.CoordinatorLayout>
